requirements.txt
openai
termcolor
python-dotenv
pillow
tenacity
matplotlib
requests
flask
tiktoken